requirements.txt
django
django-heroku
dj-database-url
django-js-asset
south
psycopg2
wrapt==1.11.*
requests
gunicorn
Pillow
html2text
setproctitle
python-dotenv
django-import-export
django-ckeditor
stripe
six
django-etc
sorl-thumbnail
django-widget-tweaks
swapper
django-sortedm2m
django-cuser
taggit
django-model-utils
django-ninja